+++ Game_Loop_Test2/src/com/company/Objects/StaticObjects/Vector.java
@@ -4,11 +4,11 @@ package com.company.Objects.StaticObjects;
  * Created by Magnus on 25-04-2016.
  */
 public class Vector {
-    public float X = 0;
-    public float Y = 0;
+    public double X = 0;
+    public double Y = 0;
 
-    public float GetLength(){
-        return (float) Math.abs(Math.sqrt( Math.pow(X, 2) + Math.pow(Y, 2)));
+    public double GetLength(){
+        return (double) Math.abs(Math.sqrt( Math.pow(X, 2) + Math.pow(Y, 2)));
     }
 
     public static Vector GetVectorByPoints(Coord p1, Coord p2){
